feat(#89): move asm related classes into separate package
