+++ src/main/java/org/eolang/jeo/representation/asm/OpcodeName.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.jeo.representation;
+package org.eolang.jeo.representation.asm;
 
 import java.lang.reflect.Field;
 import java.util.HashMap;
